Move URL+Woo constants to WooConstants
Because we already have pre-existing URL constants in there.
